% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/member.R
\name{member}
\alias{member}
\title{Get all the members of the Finnish parliament, past and present.}
\usage{
member(id = NULL, ...)
}
\arguments{
\item{id}{ID number to return information only on a MP. If no
\code{id} is provided all MPs are returned.}

\item{...}{additional arguments passed on to \code{query_kamu_api}, most
notably \code{cache} if needed.}
}
\value{
list of MPs
}
\description{
Returns all members of the Finnish parliament (MP).
}
\examples{
\dontrun{

# Get all terms
all_members <- member()

# Get a specific MP defined by ID (Jörn Donner)
jörn_donner <- member(id=30)
}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\code{\link{query_kamu_api}}
}

