% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{null_to_NA}
\alias{null_to_NA}
\title{Replace all NULLs with NAs in a data frame.}
\usage{
null_to_NA(x)
}
\arguments{
\item{x}{input data frame.
s}
}
\value{
a data frame with all NULLs replaced with NAs.
}
\description{
Replace all NULLs with NAs in a data frame.
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\keyword{internal}

