% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kamu_api.R
\name{query_kamu_api}
\alias{query_kamu_api}
\title{Query KaMu API against a given endpoint.}
\usage{
query_kamu_api(endpoint, cache, ...)
}
\arguments{
\item{endpoint}{A character URL defining the queried endpoint.}

\item{cache}{Sets the cachce mode: \code{TRUE} = Use cache if available and
save to cache, \code{FALSE} = Ignore cache if available and do not save to
cache, \code{"flush"} = Ignore cache if available and save to cache.}

\item{...}{optional query parameters.}
}
\value{
httr response object (JSON parsed to a list).
}
\description{
Function provides a single function to query KaMU API and an cache
implementation. It also defines the base URL scheme for the API. API resource
parameters are provided as arguments.
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\keyword{internal}

