% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{Get all the terms of the Finnish parliament.}
\usage{
term(id = NULL, ...)
}
\arguments{
\item{id}{ID number to return information only on a given term. If no
\code{id} is provided all terms are returned.}

\item{...}{additional argumetns passed on to \code{query_kamu_api}, most
notably \code{cache} if needed.}
}
\value{
data.frame of countries with the following columns:

\tabular{ll}{
 \code{begin} \tab Begin date. \cr
 \code{end} \tab End date. \cr
 \code{name} \tab Term name. \cr
 \code{id} \tab Term ID number. \cr
}
}
\description{
Returns all terms of the Finnish parliament 1907-2011. By default, each
term is 4 years long, but for various reasons the length may also be shorter.
}
\examples{
\dontrun{

# Get all terms
all_terms <- term()

# Get a specific term defined by ID
term_2011 <- term(id=1)
}
}
\author{
Joona Lehtomaki <joona.lehtomaki@gmail.com>
}
\seealso{
\code{\link{query_kamu_api}}
}

