% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plenary_session.R
\name{plenary_session}
\alias{plenary_session}
\title{Get all the plenary_sessions of the Finnish parliament, past and present.}
\usage{
plenary_session(id = NULL, ...)
}
\arguments{
\item{id}{ID number to return information only on a MP. If no
\code{id} is provided all MPs are returned.}

\item{...}{additional arguments passed on to \code{query_kamu_api}, most
notably \code{cache} if needed.}
}
\value{
Plenary session data in long data.frame format with the following fields:
  \itemize{
    \item{id}{Plenary session id}
    \item{field}{Data field name}
    \item{element}{Counter for elements in the given data field. Certain fields contain a list of values. This is used to distinguish between multiple list elements.}
    \item{value}{Data field value}
  }
}
\description{
Returns all plenary_sessions of the Finnish parliament (MP).
}
\examples{
\dontrun{

# Get all terms
all_plenary_sessions <- plenary_session()

# Get a specific MP defined by ID (Jorn Donner)
session <- plenary_session(id=30)
}
}
\author{
Leo Lahti <leo.lahti@iki.fi>
}
\seealso{
\code{\link{query_kamu_api}}
}

